[FEAT] 주문 완료 누르면 주문 상태 화면으로 이동 #81

+++ app/src/main/java/com/example/banchan/domain/usecase/history/GetHistoryByIdUseCase.kt
@@ -6,5 +6,5 @@ import javax.inject.Inject
 class GetHistoryByIdUseCase @Inject constructor(
     private val historyRepository: HistoryRepository
 ) {
-    operator fun invoke(id: Int) = historyRepository.getHistoryWithItemsById(id)
+    operator fun invoke(id: Long) = historyRepository.getHistoryWithItemsById(id)
 }
